--- app/houston/__init__.py
@@ -1,3 +1,0 @@
-# -*- coding: utf-8 -*-
-
-from .parameters import PatchJSONParametersWithPassword  # NOQA
