+++ flask_restplus_patched/parameters.py
@@ -3,7 +3,8 @@
 import logging
 
 from six import itervalues
-
+from flask_login import current_user
+from flask_restplus._http import HTTPStatus
 from flask_marshmallow import Schema, base_fields
 from marshmallow import validate, validates_schema, ValidationError
 
@@ -247,3 +248,54 @@ class PatchJSONParameters(Parameters):
     @classmethod
     def copy(cls, obj, field, value, state):
         raise NotImplementedError()
+
+
+# noinspection PyAbstractClass
+class PatchJSONParametersWithPassword(PatchJSONParameters):
+    """
+    Base parameters class for handling PATCH arguments according to RFC 6902 with specific handling for
+    password validation for some sensitive fields.
+    Provides test, add and remove methods.
+    """
+
+    # Some classes may require all fields to be password validated, some may require some.
+    # If the SENSITIVE_FIELDS array is left as None, all fields are password protected
+    SENSITIVE_FIELDS = None
+
+    @classmethod
+    def test(cls, obj, field, value, state):
+        from app.extensions.api import abort
+
+        if field == 'current_password':
+            if current_user.password == value:
+                state['current_password'] = value
+                return True
+            else:
+                abort(code=HTTPStatus.FORBIDDEN, message='Wrong password')
+
+        return PatchJSONParameters.test(obj, field, value, state)
+
+    @classmethod
+    def add(cls, obj, field, value, state):
+        from app.extensions.api import abort
+
+        """
+        Some or all fields require extra permissions to be changed
+        """
+        if not cls.SENSITIVE_FIELDS or field in cls.SENSITIVE_FIELDS:
+            if 'current_password' not in state:
+                abort(
+                    code=HTTPStatus.FORBIDDEN,
+                    message='Updating database requires `current_password` test operation.',
+                )
+
+    @classmethod
+    def remove(cls, obj, field, value, state):
+        from app.extensions.api import abort
+
+        if not cls.SENSITIVE_FIELDS or field in cls.SENSITIVE_FIELDS:
+            if 'current_password' not in state:
+                abort(
+                    code=HTTPStatus.FORBIDDEN,
+                    message='Updating database requires `current_password` test operation.',
+                )
